+++ src/utils/args.rs
@@ -80,7 +80,7 @@ impl From<CompressionStrategy> for CompressionLevel {
 
 /// Utility method to parse command line floats and ensure they are
 /// within the range [MIN, MAX].
-pub fn arg_in_range(arg: f32, range: std::ops::RangeInclusive<f32>) -> anyhow::Result<f32> {
+pub fn arg_in_range(arg: f64, range: std::ops::RangeInclusive<f64>) -> anyhow::Result<f64> {
     match range.contains(&arg) {
         true => Ok(arg),
         false => anyhow::bail!(
